Stop importing deprecated np.int and np.float
>>> import numpy as np
>>> np.int
<stdin>:1: DeprecationWarning: `np.int` is a deprecated alias for the builtin `int`. Use `int` by itself, which is identical in behavior, to silence this warning. If you specifically wanted the numpy scalar type, use `np.int_` here.
<class 'int'>
>>> np.float
<stdin>:1: DeprecationWarning: `np.float` is a deprecated alias for the builtin `float`. Use `float` by itself, which is identical in behavior, to silence this warning. If you specifically wanted the numpy scalar type, use `np.float_` here.
<class 'float'>
>>> np.float is float
True
>>> np.int is int
True

+++ examples/array4.py
@@ -8,7 +8,7 @@ fileh = tables.open_file(file, mode="w")
 # Get the root group
 group = fileh.root
 # Set the type codes to test
-dtypes = [np.int8, np.uint8, np.int16, np.int, np.float32, np.float]
+dtypes = [np.int8, np.uint8, np.int16, int, np.float32, float]
 i = 1
 for dtype in dtypes:
     # Create an array of dtype, with incrementally bigger ranges
